Reorganise Python libraries and paths
