+++ util/common/devices.py
